# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tallaght Community Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface Pagination {
  page: number;
  limit: number;
  total: number;
  totalPages: number;
}

export type PostStatus = (typeof PostStatus)[keyof typeof PostStatus];

export const PostStatus = {
  draft: "draft",
  held: "held",
  published: "published",
  rejected: "rejected",
} as const;

export type SubmissionStatus =
  (typeof SubmissionStatus)[keyof typeof SubmissionStatus];

export const SubmissionStatus = {
  pending: "pending",
  processing: "processing",
  processed: "processed",
  rejected: "rejected",
  failed: "failed",
} as const;

export type SubmissionSource =
  (typeof SubmissionSource)[keyof typeof SubmissionSource];

export const SubmissionSource = {
  whatsapp: "whatsapp",
  rss: "rss",
} as const;

export type Platform = (typeof Platform)[keyof typeof Platform];

export const Platform = {
  facebook: "facebook",
  instagram: "instagram",
} as const;

export type DistributionStatus =
  (typeof DistributionStatus)[keyof typeof DistributionStatus];

export const DistributionStatus = {
  pending: "pending",
  succeeded: "succeeded",
  failed: "failed",
} as const;

export interface Category {
  id: number;
  name: string;
  slug: string;
  color: string;
  createdAt: string;
}

export interface CreateCategoryRequest {
  name: string;
  slug: string;
  color?: string;
}

export interface Post {
  id: number;
  title: string;
  slug: string;
  body: string;
  excerpt?: string | null;
  headerImageUrl?: string | null;
  status: PostStatus;
  confidenceScore?: number | null;
  wordCount?: number | null;
  primaryCategoryId?: number | null;
  sourceSubmissionId?: number | null;
  isSponsored: boolean;
  isFeatured: boolean;
  publishedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreatePostRequest {
  title: string;
  slug: string;
  body: string;
  excerpt?: string;
  headerImageUrl?: string;
  status?: PostStatus;
  confidenceScore?: number;
  wordCount?: number;
  primaryCategoryId?: number;
  sourceSubmissionId?: number;
  isSponsored?: boolean;
  isFeatured?: boolean;
  categoryIds?: number[];
}

export interface UpdatePostRequest {
  title?: string;
  body?: string;
  excerpt?: string;
  headerImageUrl?: string;
  status?: PostStatus;
  confidenceScore?: number;
  primaryCategoryId?: number;
  isSponsored?: boolean;
  isFeatured?: boolean;
  publishedAt?: string;
}

export interface PostListResponse {
  posts: Post[];
  pagination: Pagination;
}

export type PostsSummaryByCategoryItem = {
  categoryId: number;
  categoryName: string;
  count: number;
};

export interface PostsSummary {
  totalPublished: number;
  totalDraft: number;
  totalHeld: number;
  todayPublished: number;
  byCategory: PostsSummaryByCategoryItem[];
}

export interface Submission {
  id: number;
  contributorId?: number | null;
  source: SubmissionSource;
  rawText?: string | null;
  mediaUrls?: string[] | null;
  voiceTranscript?: string | null;
  rssItemId?: number | null;
  status: SubmissionStatus;
  rejectionReason?: string | null;
  safetyCheckPassed?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateSubmissionRequest {
  contributorId?: number;
  source: SubmissionSource;
  rawText?: string;
  mediaUrls?: string[];
  voiceTranscript?: string;
  rssItemId?: number;
}

export interface UpdateSubmissionRequest {
  status?: SubmissionStatus;
  rejectionReason?: string;
}

export interface SubmissionListResponse {
  submissions: Submission[];
  pagination: Pagination;
}

export interface Contributor {
  id: number;
  phoneHash: string;
  displayName?: string | null;
  area?: string | null;
  bio?: string | null;
  profileImageUrl?: string | null;
  submissionCount: number;
  publishedCount: number;
  isVerified: boolean;
  isBanned: boolean;
  firstSeenAt: string;
  updatedAt: string;
}

export interface UpdateContributorRequest {
  displayName?: string;
  area?: string;
  bio?: string;
  profileImageUrl?: string;
  isVerified?: boolean;
  isBanned?: boolean;
}

export interface ContributorListResponse {
  contributors: Contributor[];
  pagination: Pagination;
}

export interface AdminUser {
  id: number;
  email: string;
  name: string;
  createdAt: string;
  lastLoginAt?: string | null;
}

export interface AdminInvitation {
  id: number;
  email: string;
  token: string;
  expiresAt: string;
  usedAt?: string | null;
  createdAt: string;
}

export interface CreateInvitationRequest {
  email: string;
}

export interface AdminLoginRequest {
  token: string;
}

export interface AdminLoginResponse {
  adminUser: AdminUser;
  sessionToken: string;
}

export interface RssFeed {
  id: number;
  name: string;
  url: string;
  isActive: boolean;
  checkIntervalMinutes: number;
  lastFetchedAt?: string | null;
  createdAt: string;
}

export interface CreateRssFeedRequest {
  name: string;
  url: string;
  checkIntervalMinutes?: number;
}

export type StatsSummaryPosts = {
  total: number;
  published: number;
  draft: number;
  held: number;
  todayPublished: number;
};

export type StatsSummarySubmissions = {
  total: number;
  pending: number;
  processed: number;
  rejected: number;
};

export type StatsSummaryContributors = {
  total: number;
  active: number;
};

export type StatsSummaryQueue = {
  pending: number;
  failed: number;
};

export interface StatsSummary {
  posts: StatsSummaryPosts;
  submissions: StatsSummarySubmissions;
  contributors: StatsSummaryContributors;
  queue: StatsSummaryQueue;
}

export interface DistributionLog {
  id: number;
  postId: number;
  platform: Platform;
  status: DistributionStatus;
  platformPostId?: string | null;
  errorMessage?: string | null;
  attemptedAt: string;
  succeededAt?: string | null;
}

export type ListPostsParams = {
  status?: PostStatus;
  categorySlug?: string;
  featured?: boolean;
  /**
   * @minimum 1
   */
  page?: number;
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
};

export type ListSubmissionsParams = {
  status?: SubmissionStatus;
  source?: SubmissionSource;
  /**
   * @minimum 1
   */
  page?: number;
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
};

export type ListContributorsParams = {
  /**
   * @minimum 1
   */
  page?: number;
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
};