# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Tallaght Community Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List posts
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/PostStatus"
        - name: categorySlug
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Paginated list of posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostListResponse"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequest"
      responses:
        "201":
          description: Post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a post by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePost
      tags: [posts]
      summary: Update a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePostRequest"
      responses:
        "200":
          description: Post updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /posts/slug/{slug}:
    get:
      operationId: getPostBySlug
      tags: [posts]
      summary: Get a post by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Post found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Categories ───────────────────────────────────────────────────────────────
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /categories/{slug}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get a category by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Category found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Submissions ──────────────────────────────────────────────────────────────
...
  /admin/invitations:
    post:
      operationId: createAdminInvitation
      tags: [admin]
      summary: Create an admin invite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvitationRequest"
      responses:
        "201":
          description: Invitation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminInvitation"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Login with an invite token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── RSS ──────────────────────────────────────────────────────────────────────
  /rss/feeds:
    get:
      operationId: listRssFeeds
      tags: [rss]
      summary: List RSS feeds
      responses:
        "200":
          description: List of RSS feeds
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RssFeed"
    post:
      operationId: createRssFeed
      tags: [rss]
      summary: Add an RSS feed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRssFeedRequest"
      responses:
        "201":
          description: Feed created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RssFeed"

  # ── Stats ────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required:
        - page
        - limit
        - total
        - totalPages

    # ── Enums ──────────────────────────────────────────────────────────────────
    PostStatus:
      type: string
      enum: [draft, held, published, rejected]
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        color:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - slug
        - color
        - createdAt
    CreateCategoryRequest:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        color:
          type: string
      required:
        - name
        - slug

    # ── Post ───────────────────────────────────────────────────────────────────
    Post:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        body:
          type: string
        excerpt:
          type: string
          nullable: true
        headerImageUrl:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/PostStatus"
        confidenceScore:
          type: number
          nullable: true
        wordCount:
          type: integer
          nullable: true
        primaryCategoryId:
          type: integer
          nullable: true
        sourceSubmissionId:
          type: integer
          nullable: true
        isSponsored:
          type: boolean
        isFeatured:
          type: boolean
        publishedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - slug
        - body
        - status
        - isSponsored
        - isFeatured
        - createdAt
        - updatedAt
    CreatePostRequest:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        body:
          type: string
        excerpt:
          type: string
        headerImageUrl:
          type: string
        status:
          $ref: "#/components/schemas/PostStatus"
        confidenceScore:
          type: number
        wordCount:
          type: integer
        primaryCategoryId:
          type: integer
        sourceSubmissionId:
          type: integer
        isSponsored:
          type: boolean
        isFeatured:
          type: boolean
        categoryIds:
          type: array
          items:
            type: integer
      required:
        - title
        - slug
        - body
    UpdatePostRequest:
      type: object
      properties:
        title:
          type: string
        body:
          type: string
        excerpt:
          type: string
        headerImageUrl:
          type: string
        status:
          $ref: "#/components/schemas/PostStatus"
        confidenceScore:
          type: number
        primaryCategoryId:
          type: integer
        isSponsored:
          type: boolean
        isFeatured:
          type: boolean
        publishedAt:
          type: string
          format: date-time
    PostListResponse:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required:
        - posts
        - pagination
...
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - email
        - name
        - createdAt
    AdminInvitation:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        token:
          type: string
        expiresAt:
          type: string
          format: date-time
        usedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - token
        - expiresAt
        - createdAt
    CreateInvitationRequest:
      type: object
      properties:
        email:
          type: string
      required:
        - email
    AdminLoginRequest:
      type: object
      properties:
        token:
          type: string
      required:
        - token
    AdminLoginResponse:
      type: object
      properties:
        adminUser:
          $ref: "#/components/schemas/AdminUser"
        sessionToken:
          type: string
      required:
        - adminUser
        - sessionToken

    # ── RSS ────────────────────────────────────────────────────────────────────
    RssFeed:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        url:
          type: string
        isActive:
          type: boolean
        checkIntervalMinutes:
          type: integer
        lastFetchedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - url
        - isActive
        - checkIntervalMinutes
        - createdAt
    CreateRssFeedRequest:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        checkIntervalMinutes:
          type: integer
      required:
        - name
        - url

    # ── Stats ──────────────────────────────────────────────────────────────────
...