# lib/api-zod/src/generated/types/createInvitationRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tallaght Community Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateInvitationRequest {
  email: string;
}