# lib/api-zod/src/generated/types/postListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tallaght Community Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Pagination } from "./pagination";
import type { Post } from "./post";

export interface PostListResponse {
  posts: Post[];
  pagination: Pagination;
}